src/llist.h: Add LLIST_{,TS}_FIND_FOREACH_CONT
In contrast to LLIST_{,TS}_FIND_FOREACH, these convenience macros search
for the first match and return successors until there is an item that
isn't matched by the filter callback. Any items beyond the first cut-off
are discarded.

Should be used when results are known to be continuous, such as
appointments and events belonging to a specific day etc.

Signed-off-by: Lukas Fleischer <calcurse@cryptocrack.de>