src/recur.c: Support recurrent multi-day appointments
Completely rewrite our inday algorithm in recur_item_inday() and be more
fine-grained. This version can deal with recurrent multi-day
appointments unless they overlap.

In case of overlapping appointments, only the last appointment that
starts before the current day is shown. We will need to rewrite the
whole recur_item_inday() interface in order to fix this - this
relatively trivial patch is only the first step.

Signed-off-by: Lukas Fleischer <calcurse@cryptocrack.de>