src/args.c: Revise app_arg()
This kills the hackish code we used to filter and display appointments
and events in non-interactive mode. From now on, the same algorithm that
is used in interactive mode (read day_store_items() for details) is
called, resulting in code that is much easier to maintain.

The resulting performance loss is rather small. Here are the run times
of `calcurse -s01/01/1902 -r18250 -D ../test/data >/dev/null` before and
after the patch was applied:

    0.30user 0.26system 0:01.22elapsed 46%CPU (0avgtext+0avgdata 1340maxresident)k
    0inputs+0outputs (0major+398minor)pagefaults 0swaps

    0.33user 0.21system 0:01.18elapsed 46%CPU (0avgtext+0avgdata 1360maxresident)k
    0inputs+0outputs (0major+395minor)pagefaults 0swaps

Note that this also fixes a few bugs that were encountered in
non-interactive mode, most notably BUG#2.

Signed-off-by: Lukas Fleischer <calcurse@cryptocrack.de>