Do not display a mark when files are auto-saved
This is kind of useless since users are not able to react to an
auto-save notification. It also causes all kinds of problems if the
status bar is in a non-standard mode (message, prompt, ...) and is prone
to race conditions if the status bar is updated by another thread at the
same time.

Signed-off-by: Lukas Fleischer <calcurse@cryptocrack.de>