run-test.c: Return failure if actual output is too short
We always terminated run-test when reached the end of the expected
output and returned success if actual output and expected output were
the same up to this point. This resulted in run-test always returning
successfully if the actual output was a prefix of the expected output,
even if it was a proper prefix.

Check if the expected output contains more data after string comparison
has finished to ensure we only return successfully if both outputs are
actually equal.

Signed-off-by: Lukas Fleischer <calcurse@cryptocrack.de>