calcurse-caldav: Multiple events in iCal file
If the server returns multiple events within an ical file, multiple
objhashes have to be saved. Because the URL is the key field, multiple
database entries are not an option.

With this commit, a semicolon separated list of hashes is created in the
database. When syncing, the list is temporarily expanded in order to
find out what needs to be synced.

The change is backwards compatible.

Signed-off-by: Stefan Hagen <github@textmail.me>