calcurse-caldav: Separate TODO and EVENT query
In order to support iCloud, we need two queries - one for EVENTs and one
for TODOs. iCloud does not allow to query without a event or todo
filter. These filters can not be combined with a logical OR.

One solution would be to use webdavs `PROPFIND` method.

In order to minimize the risk of breaking existing functionality, it's
better to stick with the webcals `REPORT` method and duplicate the
query.

Signed-off-by: Stefan Hagen <github@textmail.me>