Keybindings for UTF-8 encoded characters (final touch)
Internally characters (keys) have two representations: integers and keynames.
Keynames are characters strings, usually the name of the character. E.g. the
character A has the representations 65 and "A", and the tab character the
representations 9 and "TAB".

The function keys_int2str() turns the integer representation of a key/character
into the keyname.

For display purposes the keynames are usually confined to have displaywidth
at most three. Some curses pseudo-keys have longer keynames. E.g. the back-tab
character is "KEY_BTAB". A long keyname makes a character difficult to recognize
in the status bar menu.

The keyname of a multibyte, UTF-8 encoded character is the conventional Unicode
name of the code point. E.g. the character ü has keyname "U+00FC" because ü is
the code point 0xFC. Most of these look alike in the status bar menu.

The patch makes the keyname of a multibyte character look like that of a
singlebyte character: the character itself, i.e. the keyname of the character ü
is "ü".

The main tool is implementation of a utf8_encode() routine.