Fix appointment becoming event
You try to enter an appointment, but enter an invalid start time (by mistake).
Calcurse rejects the input. You enter the correct start time and calcurse
asks for the description, not the end time, i.e. you get an event.

The check for an event must only be performed on valid input.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>