Fix end-before-start inconsistency
Due to deficient validation, it is possible to get an inconsistent database with
an appointment that ends before it begins. Edit an existing one and specify an
end time by a date that precedes the start date, or create a new one and give a
date as end time that precedes the start. Then exit calcurse; on restart it
will report a data error and exit.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>